% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoperiod.R
\name{daylength_f95}
\alias{daylength_f95}
\title{Daylength (photoperiod) in hours (Forsythe et al. 1995)}
\usage{
daylength_f95(date, lat)
}
\arguments{
\item{date}{Date vector.}

\item{lat}{Latitude in decimal degrees (-90..90).}
}
\value{
Numeric vector of daylight length in hours.
}
\description{
Fast, dependency-free daylight length given date and latitude.
}
\references{
Forsythe, W. C. et al. (1995) Ecol. Modelling 80: 1–13.
}
