% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoperiod.R
\docType{data}
\name{.photoperiod_sites}
\alias{.photoperiod_sites}
\title{Built-in latitudes for convenience}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
.photoperiod_sites
}
\description{
Small named vector so users can supply \code{location} instead of \code{lat}.
Extend/modify as you like.
}
\keyword{internal}
