% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoperiod.R
\name{photoperiod_sites}
\alias{photoperiod_sites}
\title{List available built-in photoperiod sites}
\usage{
photoperiod_sites()
}
\value{
Named numeric vector of latitudes.
}
\description{
List available built-in photoperiod sites
}
